Move from common to oas3
