+++ packages/core/src/rules/oas3/__tests__/component-name-unique.test.ts
