+++ packages/core/src/rules/oas3/__tests__/utils/lint-document-for-test.ts
