+++ packages/core/src/rules/oas3/component-name-unique.ts
